finalize PR
- name profile data versions to be consistent with meta data versions
- move deprecated profile to profile_v2.rs
- use enum discriminants instead of manual index mapping
- add  changelog entry for new gas profiles
- update tests
