+++ integration-tests/src/tests/client/features/chunk_nodes_cache.rs
@@ -133,7 +133,8 @@ fn compare_node_counts() {
             let metadata = receipt_execution_outcome.outcome_with_id.outcome.metadata;
             match metadata {
                 ExecutionMetadata::V1 => panic!("ExecutionMetadata cannot be empty"),
-                ExecutionMetadata::V2(profile_data) => TrieNodesCount {
+                ExecutionMetadata::V2(_profile_data) => panic!("expected newest ExecutionMetadata"),
+                ExecutionMetadata::V3(profile_data) => TrieNodesCount {
                     db_reads: {
                         let cost = profile_data.get_ext_cost(ExtCosts::touching_trie_node);
                         assert_eq!(cost % touching_trie_node_cost, 0);
